+++ include/ctranslate2/layers/attention.h
@@ -6,9 +6,9 @@
 namespace ctranslate2 {
   namespace layers {
 
-    StorageView make_relative_positions(dim_t length,
-                                        dim_t max_position,
-                                        bool with_cache = false);
+    StorageView make_relative_positions(dim_t queries_length,
+                                        dim_t keys_length,
+                                        dim_t max_position);
 
     class RotaryEmbeddings;
     class Alibi;
@@ -72,15 +72,23 @@ namespace ctranslate2 {
       dim_t _maximum_relative_position;
       const float _queries_scale;
       const dim_t _num_heads_kv;
+      const bool _merge_time_and_head_dims;
       const dim_t _cache_time_dim;
+    };
+
+    enum class RotaryScalingType {
+      None = -1,
+      Linear,
     };
 
     class RotaryEmbeddings {
     public:
       RotaryEmbeddings(const dim_t dim = 0,
                        const bool interleave = true,
-                       const dim_t num_initial_positions = 2048,
-                       const float base = 10000);
+                       const RotaryScalingType scaling_type = RotaryScalingType::None,
+                       const float scaling_factor = 1,
+                       const float base = 10000,
+                       const dim_t num_initial_positions = 2048);
 
       void apply(StorageView& x, const dim_t offset = 0);
 
@@ -92,8 +100,10 @@ namespace ctranslate2 {
 
       const dim_t _dim;
       const bool _interleave;
-      const dim_t _num_initial_positions;
+      const RotaryScalingType _scaling_type;
+      const float _scaling_factor;
       const float _base;
+      const dim_t _num_initial_positions;
       const ops::Rotary _rotary_op;
 
       StorageView _sin;
@@ -103,13 +113,14 @@ namespace ctranslate2 {
 
     class Alibi {
     public:
-      Alibi(const bool use_positive_positions = false, const dim_t num_initial_positions = 2048);
+      Alibi(const bool use_positive_positions = false, const bool scale_alibi = false, const dim_t num_initial_positions = 2048);
 
-      void apply(StorageView& x);
+      void apply(StorageView& x, const float scale = 1);
 
     private:
       const bool _use_positive_positions;
       const dim_t _num_initial_positions;
+      const bool _scale_alibi;
       const ops::AlibiAdd _alibi_op;
 
       StorageView _alibi;