+++ CHANGELOG.md
@@ -4,6 +4,65 @@
 
 ### Fixes and improvements
 
+## [v3.19.0](https://github.com/OpenNMT/CTranslate2/releases/tag/v3.19.0) (2023-08-31)
+
+### Changes
+
+* Binary wheels for Python 3.7 are no longer built
+
+### New features
+
+* Build wheels for Python 3.12
+* Update the Transformers converter to support more model architectures:
+  * Falcon-RW
+  * DistilBERT
+  * Llama with linear RoPE scaling (e.g. Vicuna v1.5)
+  * Llama with a non default RoPE base period (e.g. CodeLlama)
+* Accept the token type IDs as inputs for encoder models
+* Add property `GenerationStepResult.hypothesis_id` to identify the different hypotheses when running random sampling with `num_hypotheses` > 1
+
+### Fixes and improvements
+
+* Improve performance of 8-bit models on CPU:
+  * Vectorize the GEMM output dequantization
+  * Fuse the GEMM output dequantization with bias and activation
+* Allow inputs shorter than 30 seconds in Whisper methods
+* Fix incorrect `batch_id` values passed to the callback function
+* Fix a shape error in models using both MQA and relative positions
+* Fix compilation error related to AVX512 when using GCC 7
+* Call `.detach()` on PyTorch tensors before getting the Numpy array in converters
+
+## [v3.18.0](https://github.com/OpenNMT/CTranslate2/releases/tag/v3.18.0) (2023-08-03)
+
+### Changes
+
+Converted models now uses the same floating point precision as the original models. For example, a model saved in float16 will be converted to a float16 model. Before this change, the weights were casted to float32 by default.
+
+Similarly, selecting int8 keeps non quantized weights in their original precision unless a more specific quantization type is selected:
+
+* int8_float32
+* int8_float16
+* int8_bfloat16
+
+### New features
+
+* Add property `compute_type` to model instances
+* Extend the Python class `StorageView` with additional methods and properties:
+  * `to(dtype)`
+  * `device_index`
+  * `device`
+  * `dtype`
+  * `shape`
+
+### Fixes and improvements
+
+* Update the function `get_supported_compute_types` to correctly return bfloat16 when supported
+* Update the HF Llama converter to accept extra tokens in the vocabulary
+* Fix a shape error when enabling `return_alternatives` with a model using relative positions
+* Fix a conversion error when using `torch<1.13`
+* Fix a type error when running Whisper models with the bfloat16 type
+* Update pybind11 to 2.11.1
+
 ## [v3.17.1](https://github.com/OpenNMT/CTranslate2/releases/tag/v3.17.1) (2023-07-20)
 
 ### Fixes and improvements